evaluator: move evaluator.{PassFailStatus,TestResult} to tftbase module
