+++ evaluator.py
@@ -3,7 +3,6 @@ import json
 import sys
 import yaml
 
-from dataclasses import dataclass
 from pathlib import Path
 from typing import Optional
 
@@ -11,44 +10,10 @@ import evalConfig
 import tftbase
 
 from common import dataclass_to_dict
-from common import strict_dataclass
 from logger import logger
-from tftbase import Bitrate
 from tftbase import FlowTestOutput
-from tftbase import TestMetadata
-
-
-@strict_dataclass
-@dataclass(frozen=True, kw_only=True)
-class PassFailStatus:
-    """Pass/Fail ratio and result from evaluating a full tft Flow Test result
-
-    Attributes:
-        result: boolean representing whether the test was successful (100% passing)
-        num_passed: int number of test cases passed
-        num_failed: int number of test cases failed"""
-
-    result: bool
-    num_tft_passed: int
-    num_tft_failed: int
-    num_plugin_passed: int
-    num_plugin_failed: int
-
-
-@strict_dataclass
-@dataclass(frozen=True, kw_only=True)
-class TestResult:
-    """Result of a single test case run
-
-    Attributes:
-        tft_metadata: information about which test ran
-        success: boolean representing whether the test passed or failed
-        birate_gbps: Bitrate namedtuple containing the resulting rx and tx bitrate in Gbps
-    """
-
-    tft_metadata: TestMetadata
-    success: bool
-    bitrate_gbps: Bitrate
+from tftbase import PassFailStatus
+from tftbase import TestResult
 
 
 class Evaluator:
